refactor: move class definition in src dir
